organize folder layout of Annotator related components and hooks
