+++ packages/client/src/components/advanced/Annotator/styles.ts
